requirements.txt
python-telegram-bot==20.6
qrcode
Pillow
python-dotenv
